test(mobile): split terminal mouse click and drag coverage
